refactor build process
