+++ .gitignore
@@ -1,0 +1,6 @@
+whisper
+*.o
+*.swp
+*~
+.DS_Store
+build/*
