+++ src/libwhisper.c
@@ -1,25 +1,4 @@
-#ifndef __LIBWHISPER__
-#define __LIBWHISPER__
-
-#include <stdio.h>
-#include <stdint.h>
-#include <stdlib.h>
-
-#define MAX_ARCHIVES 8
-
-struct wsp_archive_info {
-  long offset;
-  long seconds_per_point;
-  long points;
-};
-
-struct wsp_header {
-  long aggregation_type;
-  long max_retention;
-  float xff;
-  long archive_count;
-  struct wsp_archive_info archives[MAX_ARCHIVES];
-};
+#include "whisper.h"
 
 int wsp_create() {
   return 0;
@@ -27,7 +6,7 @@ int wsp_create() {
 
 int wsp_info(FILE *fd, struct wsp_header *header) {
   fpos_t original_offset;
-  struct wsp_archive_info *archive_info;
+  //struct wsp_archive_info *archive_info;
   uint8_t buf[16];
   uint32_t temp;
 
@@ -97,33 +76,3 @@ int wsp_parse_retention_data() {
 int wsp_validate_archive_list() {
   return 0;
 }
-
-// Main (tests)
-int main(int argc, char **argv) {
-  FILE *my_file;
-  struct wsp_header header;
-  char *filename = "test/mem-free.wsp";
-
-  printf("whisper info for %s\n", filename);
-  my_file = fopen(filename, "rb");
-
-  if (wsp_info(my_file, &header) == -1) {
-    return -1;
-  }
-
-  printf("> aggregation type: %lu\n", header.aggregation_type);
-  printf("> max retention:    %lu\n", header.max_retention);
-  printf("> xfilesfactor:     %f\n",  header.xff);
-  printf("> archive count:    %lu\n", header.archive_count);
-
-  for (int i=0 ; i < header.archive_count ; i++) {
-    printf("> offset:            %lu\n", header.archives[i].offset);
-    printf("> seconds per point: %lu\n", header.archives[i].seconds_per_point);
-    printf("> points:            %lu\n", header.archives[i].points);
-  }
-
-  fclose(my_file);
-  return 0;
-}
-
-#endif
