+++ src/whisper.h
@@ -1,0 +1,12 @@
+#ifndef __WHISPER__H_
+#define __WHISPER__H_
+
+#include <stdio.h>
+#include <stdint.h>
+#include <stdlib.h>
+
+#define MAX_ARCHIVES 8
+
+#include "libwhisper.h"
+
+#endif
